finally, consensus clustering...
this is not efficient, but necessary for publication level clustering. The mose important thing is to set a way to determine resolution and cluster number based on louvain/leiden's own property. This is a way to decrease or qualify randomness, but this is not magic. Ironically randomness is very small when you have good data ;)